feat(platform): consolidate unified GKE MCP control plane and harden SRE playbooks
- Implement unified stdio Platform MCP Server in platform_mcp_server.py to natively manage operator agent provision, deprovision, completions queries, and fleet listings.
- Consolidate dynamic GKE IPAM, GCP region validation, temporary manifest generation, and kubectl subprocess execution natively inside Python, completely eliminating legacy skill helper scripts.
- Refactor agent playbooks in SOUL.md to remove redundant/overlapping playbooks and conceptually decouple all technical tool names and GKE infrastructure configurations.
- Enforce clean, human-readable status reporting guidelines inside the agent's core behavioral rules.
- Purge all manual '/opt/data/operator_agents.jsonl' database references from SRE governance SOP files, refactoring them to call the native list_operators MCP tool.
- Align GKE pod SecurityContext 'fsGroup' GID to 10000 inside deployment.yaml to resolve projected Service Account token read permission denied crashes.
- Refine Dockerfile to natively merge shared and platform config.yaml files at build-time using a clean, lightweight Python venv PyYAML merger command.
- Harden MCP server subprocess exception handling by unmasking CalledProcessError stderr output to prevent swallowed GKE/kubectl CLI errors.

TAG=agy
CONV=53a7d6e4-fa77-4f6e-a794-905c88fb36f9
