+++ agents/platform/Dockerfile
@@ -3,7 +3,6 @@ FROM nousresearch/hermes-agent:sha-bec2250d2c8349fc85201bcd1aa39bcaa766a555
 USER root
 
 # Install system utilities, gcloud CLI, GKE auth plugin, patch, and kubectl
-# Re-added patch utility as it is required by the entrypoint.patch step
 RUN apt-get update && apt-get install -y apt-transport-https ca-certificates gnupg curl patch && \
     curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | gpg --dearmor -o /usr/share/keyrings/cloud.google.gpg && \
     echo "deb [signed-by=/usr/share/keyrings/cloud.google.gpg] https://packages.cloud.google.com/apt cloud-sdk main" | tee -a /etc/apt/sources.list.d/google-cloud-sdk.list && \
@@ -14,18 +13,24 @@ RUN apt-get update && apt-get install -y apt-transport-https ca-certificates gnu
 RUN curl -fsSL -o /usr/local/bin/kubectl "https://dl.k8s.io/release/v1.30.0/bin/linux/amd64/kubectl" && \
     chmod +x /usr/local/bin/kubectl
 
-# Install Google Chat dependencies
-RUN uv pip install google-cloud-pubsub google-api-python-client google-auth google-auth-oauthlib
+# Create defaults directory structure
+RUN mkdir -p /opt/defaults/scripts
 
-# Create defaults directory
-RUN mkdir -p /opt/defaults
-
-# Copy identity from templates to defaults
+# Copy SOUL.md identity
 COPY agents/platform/SOUL.md /opt/defaults/SOUL.md
 
-# Setup defaults
+# 1. Copy shared defaults (contains config.yaml with standard 'model' block)
 COPY agents/shared/defaults/ /opt/defaults/
-COPY agents/platform/defaults/ /opt/defaults/
+
+# 2. Copy platform defaults configuration as temp file (contains 'mcp_servers' block)
+COPY agents/platform/defaults/config.yaml /tmp/platform-config.yaml
+
+# 3. Copy platform scripts directly into defaults scripts folder
+COPY agents/platform/defaults/scripts/ /opt/defaults/scripts/
+
+# 4. Dynamically merge config.yamls using the virtual environment's python binary (which has PyYAML installed)
+RUN /opt/hermes/.venv/bin/python3 -c "import yaml; s=yaml.safe_load(open('/opt/defaults/config.yaml')) or {}; p=yaml.safe_load(open('/tmp/platform-config.yaml')) or {}; s.update(p); yaml.safe_dump(s, open('/opt/defaults/config.yaml', 'w'))" && rm /tmp/platform-config.yaml
+
 RUN chmod -R a+rX /opt/defaults
 
 # Copy and apply patch
