+++ agents/platform/defaults/governance/obtainability_audit_sop.md
@@ -7,7 +7,7 @@
 ## Execution Checklist
 
 ### 1. Auditing Target Fleet
-*   Read `/opt/data/operator_agents.jsonl` to identify active GKE clusters.
+*   Call the native MCP tool `mcp_platform_control_list_operators` to retrieve the active GKE clusters list.
 
 ### 2. Obtainability & Rigidity Auditing Rules
 For each GKE cluster, query the Operator Agent to check workload configuration rigidity:
